requirements.txt
crewai
html2text 
playwright 
python-dotenv
selenium
crewai-tools
streamlit